+++ tests/components/kostal_plenticore/test_number.py
@@ -5,10 +5,9 @@ from unittest.mock import AsyncMock, MagicMock
 from kostal.plenticore import SettingsData
 import pytest
 
-from homeassistant.components.kostal_plenticore.const import (
+from homeassistant.components.kostal_plenticore.number import (
     PlenticoreNumberEntityDescription,
 )
-from homeassistant.components.kostal_plenticore.number import PlenticoreDataNumber
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_registry import async_get
 
@@ -90,108 +89,3 @@ async def test_setup_no_entries(
     ent_reg = async_get(hass)
     assert ent_reg.async_get("number.scb_battery_min_soc") is None
     assert ent_reg.async_get("number.scb_battery_min_home_consumption") is None
-
-
-def test_number_returns_value_if_available(
-    mock_coordinator: MagicMock,
-    mock_number_description: PlenticoreNumberEntityDescription,
-    mock_setting_data: SettingsData,
-):
-    """Test if value property on PlenticoreDataNumber returns an int if available."""
-
-    mock_coordinator.data = {"moduleid": {"dataid": "42"}}
-
-    entity = PlenticoreDataNumber(
-        mock_coordinator, "42", "scb", None, mock_number_description, mock_setting_data
-    )
-
-    assert entity.value == 42
-    assert type(entity.value) == int
-
-
-def test_number_returns_none_if_unavailable(
-    mock_coordinator: MagicMock,
-    mock_number_description: PlenticoreNumberEntityDescription,
-    mock_setting_data: SettingsData,
-):
-    """Test if value property on PlenticoreDataNumber returns none if unavailable."""
-
-    mock_coordinator.data = {}  # makes entity not available
-
-    entity = PlenticoreDataNumber(
-        mock_coordinator, "42", "scb", None, mock_number_description, mock_setting_data
-    )
-
-    assert entity.value is None
-
-
-async def test_set_value(
-    mock_coordinator: MagicMock,
-    mock_number_description: PlenticoreNumberEntityDescription,
-    mock_setting_data: SettingsData,
-):
-    """Test if set value calls coordinator with new value."""
-
-    entity = PlenticoreDataNumber(
-        mock_coordinator, "42", "scb", None, mock_number_description, mock_setting_data
-    )
-
-    await entity.async_set_native_value(42)
-
-    mock_coordinator.async_write_data.assert_called_once_with(
-        "moduleid", {"dataid": "42"}
-    )
-    mock_coordinator.async_refresh.assert_called_once()
-
-
-async def test_minmax_overwrite(
-    mock_coordinator: MagicMock,
-    mock_number_description: PlenticoreNumberEntityDescription,
-):
-    """Test if min/max value is overwritten from retrieved settings data."""
-
-    setting_data = SettingsData(
-        {
-            "min": "5",
-            "max": "100",
-        }
-    )
-
-    entity = PlenticoreDataNumber(
-        mock_coordinator, "42", "scb", None, mock_number_description, setting_data
-    )
-
-    assert entity.min_value == 5
-    assert entity.max_value == 100
-
-
-async def test_added_to_hass(
-    mock_coordinator: MagicMock,
-    mock_number_description: PlenticoreNumberEntityDescription,
-    mock_setting_data: SettingsData,
-):
-    """Test if coordinator starts fetching after added to hass."""
-
-    entity = PlenticoreDataNumber(
-        mock_coordinator, "42", "scb", None, mock_number_description, mock_setting_data
-    )
-
-    await entity.async_added_to_hass()
-
-    mock_coordinator.start_fetch_data.assert_called_once_with("moduleid", "dataid")
-
-
-async def test_remove_from_hass(
-    mock_coordinator: MagicMock,
-    mock_number_description: PlenticoreNumberEntityDescription,
-    mock_setting_data: SettingsData,
-):
-    """Test if coordinator stops fetching after remove from hass."""
-
-    entity = PlenticoreDataNumber(
-        mock_coordinator, "42", "scb", None, mock_number_description, mock_setting_data
-    )
-
-    await entity.async_will_remove_from_hass()
-
-    mock_coordinator.stop_fetch_data.assert_called_once_with("moduleid", "dataid")
